use vitest for eslint tests
