+++ packages/eslint-plugin/src/tests/require-root-class.spec.cjs
@@ -1,105 +1,109 @@
 "use strict";
 
-const { RuleTester } = require("eslint");
+const { runClassic } = require("eslint-vitest-rule-tester");
 const rule = require("../rules/require-root-class.cjs");
 
-const tester = new RuleTester({
-  languageOptions: {
-    parser: require("vue-eslint-parser"),
-    ecmaVersion: 2020,
-    sourceType: "module",
-  },
-});
-tester.run("require-onyx-component-class", rule, {
-  valid: [
-    {
-      code: `
+runClassic(
+  "require-onyx-component-class",
+  rule,
+  {
+    valid: [
+      {
+        code: `
         <template>
           <div class="onyx-component">
             <p>Valid component</p>
           </div>
         </template>
       `,
-    },
-    {
-      code: `
+      },
+      {
+        code: `
         <template>
           <div :class="'onyx-component'">
             <p>Dynamic class</p>
           </div>
         </template>
       `,
-    },
-    {
-      code: `
+      },
+      {
+        code: `
         <template>
           <div :class="['class1', 'onyx-component']">
             <p>Array class binding</p>
           </div>
         </template>
       `,
-    },
-    {
-      code: `
+      },
+      {
+        code: `
         <template>
           <div :class="{ 'onyx-component': true, 'other-class': false }">
             <p>Object class binding</p>
           </div>
         </template>
       `,
-    },
-    {
-      code: `
+      },
+      {
+        code: `
         <template>
           <onyx-custom-element>
             <p>Custom Onyx element</p>
           </onyx-custom-element>
         </template>
       `,
-    },
-  ],
-  invalid: [
-    {
-      code: `
+      },
+    ],
+    invalid: [
+      {
+        code: `
         <template>
           <div>
             <p>Missing class</p>
           </div>
         </template>
       `,
-      errors: [
-        {
-          messageId: "missingClass",
-        },
-      ],
-    },
-    {
-      code: `
+        errors: [
+          {
+            messageId: "missingClass",
+          },
+        ],
+      },
+      {
+        code: `
         <template>
           <div :class="['class1', 'class2']">
             <p>Array binding without onyx-component</p>
           </div>
         </template>
       `,
-      errors: [
-        {
-          messageId: "missingClass",
-        },
-      ],
-    },
-    {
-      code: `
+        errors: [
+          {
+            messageId: "missingClass",
+          },
+        ],
+      },
+      {
+        code: `
         <template>
           <div :class="{ 'other-class': true }">
             <p>Object binding without onyx-component</p>
           </div>
         </template>
       `,
-      errors: [
-        {
-          messageId: "missingClass",
-        },
-      ],
+        errors: [
+          {
+            messageId: "missingClass",
+          },
+        ],
+      },
+    ],
+  },
+  {
+    languageOptions: {
+      parser: require("vue-eslint-parser"),
+      ecmaVersion: 2020,
+      sourceType: "module",
     },
-  ],
-});
+  },
+);
